샘플링을 통해 평균 FPS 구하는 부분 출력 오류 수정 완료

+++ CoreEngine/CTimer.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <chrono>
 
-const ULONG MAX_SAMPLE_COUNT = 50;
+const int32 MAX_SAMPLE_COUNT = 50;
 const int32 DEAD_RECORNING_TPS = 30;
